pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fhir-synth"
version = "0.1.0"
description = "Deterministic FHIR R4 synthetic data generator with LLM-assisted planning"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Alvin Henrick", email = "share.code@aol.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer[all]>=0.9.0",
    "python-dateutil>=2.8",
    "fhir.resources>=7.0",
    "litellm>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-xdist>=3.0",
    "freezegun>=1.2",
    "ruff>=0.1",
    "mypy>=1.0",
    "types-pyyaml",
    "types-python-dateutil",
]

[project.scripts]
fhir-synth = "fhir_synth.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/fhir_synth"]

[tool.hatch.envs.default]
installer = "uv"
dependencies = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-xdist>=3.0",
    "freezegun>=1.2",
]

[tool.hatch.envs.default.scripts]
test = "pytest {args}"
cov = "pytest --cov=fhir_synth --cov-report=term-missing {args}"
test-fast = "pytest -n auto {args}"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B904",  # raise from exception (optional in our case)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__
"tests/*" = ["B011"]  # assert False

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "llm: marks tests that require LLM providers",
]
